eslint and other stuff
