+++ .eslintrc.js
@@ -1,0 +1,30 @@
+module.exports = {
+  env: {
+    browser: false,
+    commonjs: true,
+    es2021: true,
+    node: true
+  },
+  extends: 'eslint:recommended',
+  parserOptions: {
+    ecmaVersion: 12
+  },
+  rules: {
+    'indent': [
+      'error',
+      2
+    ],
+    'linebreak-style': [
+      'error',
+      'unix'
+    ],
+    'quotes': [
+      'error',
+      'single'
+    ],
+    'semi': [
+      'error',
+      'always'
+    ]
+  }
+};
