--- actions/index.js
@@ -1,11 +1,0 @@
-const { getImage, getImages, getLatestImage, getLatestImages, getRandomImage } = require('./get-image');
-const putImage = require('./put-image');
-
-module.exports = {
-  getImage,
-  getImages,
-  getLatestImage,
-  getLatestImages,
-  getRandomImage,
-  putImage,
-};
