--- events/discord-ready.js
@@ -1,3 +1,0 @@
-module.exports = () => {
-  console.log('Logged in to Discord');
-}
