--- events/index.js
@@ -1,7 +1,0 @@
-const discordMessage = require('./discord-message');
-const discordReady = require('./discord-ready');
-
-module.exports = {
-  discordMessage,
-  discordReady,
-};
