+++ server/actions/index.js
@@ -1,0 +1,11 @@
+const { getImage, getImages, getLatestImage, getLatestImages, getRandomImage } = require('./get-image');
+const putImage = require('./put-image');
+
+module.exports = {
+  getImage,
+  getImages,
+  getLatestImage,
+  getLatestImages,
+  getRandomImage,
+  putImage,
+};
