+++ server/events/discord-ready.js
@@ -1,0 +1,3 @@
+module.exports = () => {
+  console.log('Logged in to Discord');
+};
