+++ server/events/index.js
@@ -1,0 +1,7 @@
+const discordMessage = require('./discord-message');
+const discordReady = require('./discord-ready');
+
+module.exports = {
+  discordMessage,
+  discordReady,
+};
